+++ src/socketio/mod.rs
@@ -5,3 +5,8 @@ mod packet;
 /// could be sent or received.
 pub mod payload;
 pub(crate) mod socket;
+pub mod socket_builder;
+
+// For ease of access
+pub type SocketBuilder = socket_builder::SocketBuilder;
+pub type Socket = socket::SocketIOSocket;
